requirements.txt
ultralytics==8.2.50
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
opencv-python==4.10.0.84
